refactor(model)!: Move ClearlyDefined and PURL code to separate files
The respective coordinate mapping functions deserve individual files.
This also helps to semantically compare ClearlyDefined and PURL mapping.
