+++ plugins/reporters/asciidoc/src/main/resources/templates/asciidoc/vulnerability_report.ftl
@@ -17,7 +17,7 @@
     License-Filename: LICENSE
 --]
 
-[#assign ModelExtensions = statics['org.ossreviewtoolkit.model.utils.ExtensionsKt']]
+[#assign PurlUtils = statics['org.ossreviewtoolkit.model.utils.PurlUtilsKt']]
 
 :publisher: OSS Review Toolkit
 [#assign now = .now]
@@ -46,7 +46,7 @@ lack relevant vulnerabilities. Further details about the issues that occurred ca
 == Packages
 [#assign advisorResults = helper.advisorResultsWithVulnerabilities()]
 [#list advisorResults as id, results]
-=== ${ModelExtensions.toPurl(id)}
+=== ${PurlUtils.toPurl(id)}
 
 [#list results as result]
 
@@ -85,7 +85,7 @@ occurred when requesting vulnerability information for the single packages.
 [#list advisorResultsWithErrors as id, results]
 === ${id.name}
 
-${ModelExtensions.toPurl(id)}
+${PurlUtils.toPurl(id)}
 
 [#list results as result]
